+++ web/src/component.dart
@@ -49,3 +49,5 @@ part 'component/widget/hidden_model.dart';
 part 'component/widget/html_model.dart';
 part 'component/widget/hyperlink_model.dart';
 part 'component/widget/image_model.dart';
+part 'component/widget/inline_html_model.dart';
+part 'component/widget/inline_hyperlink_model.dart';